+++ src/main/webapp/WEB-INF/views/guestdelluna/myPayList.jsp
@@ -85,7 +85,7 @@
 <div id="wrap" style="margin-top: 3rem;">
     <div class="menuL" style="margin-top: 6.5rem;">
         <ul>
-            <li><a href="/mypage/reviews.page" style="color: black;">후기</a></li>
+            <li><a href="/my-page/reviews.page" style="color: black;">후기</a></li>
             <c:if test="${memberCode == memberDto.memberCode}">
                 <li><a href="${root}/guestdelluna/memberUpdate.do" style="color: black;">회원수정</a></li>
                 <li><a href="${root}/manage/points.page" style="color: black;">포인트관리</a></li>
