+++ src/main/webapp/WEB-INF/views/guestdelluna/zzimlist.jsp
@@ -90,7 +90,7 @@
 <div id="wrap" style="margin-top: 3rem;">
     <div class="menuL" style="margin-top: 6.5rem;">
         <ul>
-            <li><a href="/mypage/reviews.page"
+            <li><a href="/my-page/reviews.page"
                    style="color: black;">후기</a></li>
             <c:if test="${memberCode == memberDto.memberCode}">
                 <li><a href="${root}/guestdelluna/memberUpdate.do"
