+++ src/main/webapp/WEB-INF/views/host/houseManagement.jsp
@@ -47,7 +47,7 @@
     <div class="menuL">
         <ul>
             <li>
-                <a href="/mypage/reviews.page">후기</a>
+                <a href="/my-page/reviews.page">후기</a>
             </li>
             <li>
                 <a href="${root}/guestdelluna/memberUpdate.do">회원수정</a>
